Add selected sections to outline

Add selected sections to outline

Usage

add_base_section(x)

Arguments

x

inherited 'custom_sections' parameter from create_template

Value

Call and copy the sections in the package templates to create an outline for a stock assessment

Examples

add_base_section(c("executive summary", "modeling approach", "results"))
#> [[1]]
#> [1] "executive_summary.qmd"
#> 
#> [[2]]
#> [1] "modeling_approach.qmd"
#> 
#> [[3]]
#> [1] "results.qmd"
#>