Convert Output

Format stock assessment output files to a standardized format.

Usage

convert_output(
  output.file = NULL,
  outdir = NULL,
  model = NULL,
  fleet_names = NULL
)

Arguments

output.file

name of the file containing assessment model output. This is the Report.sso file for SS3, the rdat file for BAM, the...

outdir

output directory folder

model

assessment model used in evaluation; "ss3", "bam", "asap", "fims", "amak", "ms-java", "wham", "mas", "asap"

fleet_names

Names of fleets in the assessment model as shortened in the output file, required for transforming BAM model output

Value

A reformatted and standardized version of assessment model results for application in building a stock assessment reports and to easily adapt results among regional assessments. The resulting object is simply a transformed and machine readable version of a model output file. There are 2 options for adding data to the function. (1) Add the full path with the file name in output.file or (2) output.file is the file name and outdir is the path to the file without a trailing forward slash.

Author

Samantha Schiano