Convert Output

Format stock assessment output files to a standardized format.

convert_output(
  output_file = NULL,
  outdir = getwd(),
  model = NULL,
  fleet_names = NULL,
  file_save = FALSE,
  savedir = getwd(),
  save_name = "std_model_output"
)

Arguments

output_file

Assessment model output file (e.g., the Report.sso file for SS3, the rdat file for BAM, etc.)

outdir

Directory of the assessment model output file. Defaults to working directory.

model

Assessment model used in evaluation ("ss3", "bam", "asap", "fims", "amak", "ms-java", "wham", "mas").

fleet_names

Names of fleets in the assessment model as shortened in the output file. If fleet names are not properly read, then indicate the fleets names as an acronym in a vector

file_save

TRUE/FALSE; Save the formatted object rather than calling the function and adding the formatted object to the global environment? Default is false.

savedir

Directory to save the converted output file. Defaults to working directory.

save_name

Name of the converted output file (do not use spaces).

Value

A reformatted and standardized version of assessment model results for application in building a stock assessment reports and to easily adapt results among regional assessments. The resulting object is simply a transformed and machine readable version of a model output file. There are 2 options for adding data to the function. (1) Add the full path with the file name in output.file or (2) output.file is the file name and outdir is the path to the file without a trailing forward slash.

Author

Samantha Schiano

Examples

if (FALSE) { # \dontrun{
convert_output(
output_file = "Report.sso",
outdir = "data",
model = "ss3",
savedir = getwd(),
save_name = "converted_model_output")
} # }