Generate Citation for Stock Assessment Report

Generate Citation for Stock Assessment Report

Usage

create_citation(author = NULL, title = NULL, year = NULL, office = NULL)

Arguments

author

author list- inherited from template function is using

title

title of the report - inherited from template function if using

year

year of the report for citation purposes

office

NOAA line office that the report is being processed out of

Value

Generate a citation for use in publications and other references for the stock assessment report generated in the ASAR package.

Examples

create_citation(
  title = "SA Report for Jellyfish",
  author = c("John Snow", "Danny Phantom", "Patrick Star"),
  year = 2024, office = "NEFSC"
)
#> [1] "{{< pagebreak >}} \n\nPlease cite this publication as \n\nSnow, J., D. Phantom, P. Star. 2024. SA Report for Jellyfish. NOAA Fisheries Science Center, Juneau, AK. "