Create Quarto Document of Tables

Create Quarto Document of Tables

Usage

create_tables_doc(
  resdir = NULL,
  model_results = NULL,
  model = c("SS3", "BAM", "ASAP", "AMAK", "WHAM"),
  subdir = NULL,
  include_all = TRUE
)

Arguments

resdir

directory where the results file is located

model_results

name of the results file of assessment output

model

stock assessment model

subdir

subdirectory where the assessment report template is being stored

include_all

include all default tables for a stock assessment report

Value

Create a quarto document as part of a stock assessment outline with pre-loaded R chunk adding the stock assessment tables from the nmfs-ost/satf R package