Write Stock Assessment Title

Write Stock Assessment Title

Usage

create_title(
  office = NULL,
  species = NULL,
  region = NULL,
  year = NULL,
  complex = NULL,
  type = NULL,
  spp_latin = NULL
)

Arguments

office

Fishery science center performing the assessment

species

Target species

region

Region where the stock is being assessed if applicable. Please be explicit in the name and follow the preferred convention for the management organization.

year

Year conducting the assessment report

complex

TRUE/FALSE is this a species complex?

type

Stock assessment type; usually inherited from the template

spp_latin

Latin name for the target species

Examples

create_title(
  office = "SEFSC", species = "Red Snapper", region = "South Atlantic",
  year = 2024, type = "SAR", spp_latin = "Lutjanus campechanus"
)
#> [1] "SEDAR XX Assessment Report for Red Snapper (Lutjanus campechanus) in the South Atlantic in 2024"